move globFilesSafely() to basic module
