--- pyBioinfo/pyBioinfo_modules/basic/files.py
@@ -1,9 +1,0 @@
-from pathlib import Path
-
-def findStartLine(csvFile: Path) -> int:
-    with csvFile.open('r') as file:
-        for i, line in enumerate(file):
-            if not line.strip()[0] == '#':
-                if line != '':
-                    return i
-        return 0
